dom/tv/test/test_tv_non_permitted_app.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Non-Permitted Application for TV API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

"use strict";

var gHostedManifestURL = 'http://test/tests/dom/tv/test/file_app.sjs?testToken=file_tv_non_permitted_app.html&template=file_app.template.webapp';
var gApp;

function cbError(e) {
  ok(false, "Error callback invoked: " + this.error.name);
  SimpleTest.finish();
}

function installApp() {
  var request = navigator.mozApps.install(gHostedManifestURL);
  request.onerror = cbError;
  request.onsuccess = function() {
    gApp = request.result;
    runTest();
  }
}

function uninstallApp() {
  var request = navigator.mozApps.mgmt.uninstall(gApp);
  request.onerror = cbError;
  request.onsuccess = function() {
    // All done.
    info("All done");
    runTest();
  }
}

function testApp() {
  var ifr = document.createElement('iframe');
  ifr.setAttribute('mozbrowser', 'true');
  ifr.setAttribute('mozapp', gApp.manifestURL);
  ifr.setAttribute('src', gApp.manifest.launch_path);
  var domParent = document.getElementById('content');

  // Set us up to listen for messages from the app.
  var listener = function(e) {
    var message = e.detail.message;
    if (/^OK/.exec(message)) {
      ok(true, "Message from app: " + message);
    } else if (/KO/.exec(message)) {
      ok(false, "Message from app: " + message);
    } else if (/DONE/.exec(message)) {
      ok(true, "Messaging from app complete");
      ifr.removeEventListener('mozbrowsershowmodalprompt', listener);
      domParent.removeChild(ifr);
      runTest();
    }
  }

  // This event is triggered when the app calls "alert".
  ifr.addEventListener('mozbrowsershowmodalprompt', listener, false);
  domParent.appendChild(ifr);
}

var tests = [
  // Installing the app
  installApp,

  // Run tests in app
  testApp,

  // Uninstall the app
  uninstallApp
];

function runTest() {
  if (!tests.length) {
    SimpleTest.finish();
    return;
  }

  var test = tests.shift();
  test();
}

SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv({"set": [["dom.tv.enabled", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{ "type": "tv", "allow": true, "context": document },
     { "type": "browser", "allow": true, "context": document },
     { "type": "embed-apps", "allow": true, "context": document },
     { "type": "webapps-manage", "allow": true, "context": document }],
    function(){
      SpecialPowers.setAllAppsLaunchable(true);
      SpecialPowers.setBoolPref("dom.mozBrowserFramesEnabled", true);
      // No confirmation needed when an app is installed
      SpecialPowers.autoConfirmAppInstall(() => {
        SpecialPowers.autoConfirmAppUninstall(runTest);
      });
    });
});

</script>
</pre>
</body>
</html>
